+++ index.d.ts
@@ -98,6 +98,8 @@ export interface LivePreviewConfigWithPreviewToken {
 export interface LivePreviewQuery {
     live_preview: string
     content_type_uid: string
+    preview_timestamp: string
+    release_id: string
 }
 
 export interface RetryDelayOption {
@@ -170,6 +172,7 @@ export function Stack(config: Config): Stack;
  */
 export function Stack(api_key: string, access_token: string, environment_name: string, region?: string, fetchOptions?: FetchOptions): Stack;
 
+export function updateAssetURL(entry: object): object;
 export class ContentType {
     constructor();
     content_type_uid: string